modify: use connector id from rainbow kit wallets
