+++ components/CryptoIcon.tsx
@@ -12,9 +12,9 @@ export type CryptoIconType =
   | 'binance'
   | 'bnb'
   | 'bsc'
-  | 'coinbaseWallet'
+  | 'coinbase'
   | 'ether'
-  | 'trustWallet'
+  | 'trust'
   | 'metaMask'
   | 'walletConnect'
   | 'wncg'
